fix answer voting

+++ public/javascripts/question.js
@@ -29,10 +29,14 @@ window.addEventListener("DOMContentLoaded", async (event) => {
   downVoteQ.addEventListener("click", (e) => questionVote(2, e.target.id));
 
   const upVoteA = document.querySelectorAll(".answer-upvote-arrow");
-    const downVoteA = document.querySelectorAll(".answer-downvote-arrow");;
-  upVoteA.forEach((upVoteButton) => {
-    upVoteButton.addEventListener("click", (e) => answerVote(1, e.target.id));
-  });
+  const downVoteA = document.querySelectorAll(".answer-downvote-arrow");
+  upVoteA.forEach(upVoteButton => {
+          upVoteButton.addEventListener("click", (e) => answerVote(1, e.target.id));
+  })
+
+  downVoteA.forEach(downVoteButton => {
+      downVoteButton.addEventListener("click", (e) => answerVote(2, e.target.id));
+  })
 
   if (editButton) {
     editButton.addEventListener("click", (event) => {
@@ -127,41 +131,44 @@ window.addEventListener("DOMContentLoaded", async (event) => {
       }
     });
   }
-  answerQuestionButton.addEventListener("click", async (e) => {
-    e.preventDefault();
 
-    const id = e.target.id;
-
-    const answerId = await postAnswer(`answerquestion/${id}`, answerTextBox);
-
-
-    const newAnswerDiv = document.createElement("div");
-
-    newAnswerDiv.setAttribute("id", `answer-${id}-div`);
-    newAnswerDiv.setAttribute("class", "answer-div");
-    const newAnswerButtons = document.createElement("div");
-
-    const newAnswerDelete = document.createElement("BUTTON");
-    newAnswerDelete.innerHTML = "Delete answer";
-    newAnswerDelete.setAttribute("class", "button");
-
-    newAnswerDelete.addEventListener("click", async (e) => {
+  if (answerQuestionButton) {
+    answerQuestionButton.addEventListener("click", async (e) => {
       e.preventDefault();
-      removeDiv(id);
 
-      await deleteItem("Answer", "answers", answerId);
+      const id = e.target.id;
+
+      const answerId = await postAnswer(`answerquestion/${id}`, answerTextBox);
+
+
+      const newAnswerDiv = document.createElement("div");
+
+      newAnswerDiv.setAttribute("id", `answer-${id}-div`);
+      newAnswerDiv.setAttribute("class", "answer-div");
+      const newAnswerButtons = document.createElement("div");
+
+      const newAnswerDelete = document.createElement("BUTTON");
+      newAnswerDelete.innerHTML = "Delete answer";
+      newAnswerDelete.setAttribute("class", "button");
+
+      newAnswerDelete.addEventListener("click", async (e) => {
+        e.preventDefault();
+        removeDiv(id);
+
+        await deleteItem("Answer", "answers", answerId);
+      });
+
+      newAnswerButtons.appendChild(newAnswerDelete);
+
+      const newAnswerText = document.createElement("div");
+      newAnswerText.innerHTML = answerTextBox.value;
+      answerTextBox.value = "";
+      answersDiv.appendChild(newAnswerDiv);
+      newAnswerDiv.appendChild(newAnswerText);
+      newAnswerDiv.appendChild(newAnswerButtons);
+      document.querySelector(".answerQuestionForm").style.display = "none";
     });
-
-    newAnswerButtons.appendChild(newAnswerDelete);
-
-    const newAnswerText = document.createElement("div");
-    newAnswerText.innerHTML = answerTextBox.value;
-    answerTextBox.value = "";
-    answersDiv.appendChild(newAnswerDiv);
-    newAnswerDiv.appendChild(newAnswerText);
-    newAnswerDiv.appendChild(newAnswerButtons);
-    document.querySelector(".answerQuestionForm").style.display = "none";
-  });
+  }
 
   if (deleteQuestion) {
     deleteQuestion.addEventListener("click", async (e) => {
@@ -173,17 +180,6 @@ window.addEventListener("DOMContentLoaded", async (event) => {
 
   if (deleteAnswers.length) {
     // Use a for loop to add an event listener to each answer div
-
-
-    upVoteA.forEach((upVoteButton) => {
-      upVoteButton.addEventListener("click", (e) => answerVote(1, e.target.id));
-    });
-    downVoteA.forEach((downVoteButton) => {
-      downVoteButton.addEventListener("click", (e) =>
-        answerVote(2, e.target.id)
-      );
-    });
-
     for (let i = 0; i < deleteAnswers.length; i++) {
       deleteAnswers[i].addEventListener("click", async (e) => {
         const target = e.target;
@@ -195,6 +191,9 @@ window.addEventListener("DOMContentLoaded", async (event) => {
       });
     }
   }
+
+
+
 });
 
 const removeDiv = function (id) {
@@ -299,66 +298,67 @@ async function questionVote(upOrDownCode, questionId) {
 }
 
 async function answerVote(upOrDownCode, answerId) {
-  const score = document.querySelector(
-    `${answerId}.question-page-answer-score`
-  );
-  const upvoteButton = document.querySelector(`${answerId}.fa-plus-circle`);
-  const downvoteButton = document.querySelector(`${answerId}.fa-minus-circle`);
 
-  //answerId has an 'A' prefix that needs to be removed
-  answerId = answerId.slice(1);
+    answerId = answerId.slice(1);
+    answerId = answerId.split('-')[0];
+    const score = document.getElementById(`A${answerId}-score`)
+    const upvoteButton = document.getElementById(`A${answerId}-up`);
+    const downvoteButton = document.getElementById(`A${answerId}-down`);
 
-  // // previously voted
-  // if (upOrDownCode === 1 && upVoteQ.classList.contains("upvoted-arrow")) {
-  //     try {
-  //         await fetch(`http://localhost:8080/answers/${answerId}/vote`, {
-  //             method: `DELETE`,
-  //         });
-  //     } catch (err) {
-  //         console.log("question vote error", err);
-  //     }
-  //     upVoteQ.classList.remove("upvoted-arrow");
-  //     score.innerText--;
-  //     return;
-  // }
 
-  // // previously voted
-  // if (upOrDownCode === 2 && downVoteQ.classList.contains("downvoted-arrow")) {
-  //     try {
-  //         await fetch(`http://localhost:8080/answers/${answerId}/vote`, {
-  //             method: `DELETE`,
-  //         });
-  //     } catch (err) {
-  //         console.log("question vote error", err);
-  //     }
-  //     downVoteQ.classList.remove("downvoted-arrow");
-  //     score.innerText++;
-  //     return;
-  // }
-
-  // fresh vote
-  try {
-    await fetch(`http://localhost:8080/answers/${answerId}/vote`, {
-      method: `DELETE`,
-    });
-    await fetch(
-      `http://localhost:8080/answers/${answerId}/vote/${upOrDownCode}`,
-      { method: `POST` }
-    );
-
-    if (upOrDownCode === 1) {
-      //upvote
-      if (downvoteButton.classList.contains("downvoted-arrow"))
-        score.innerText++;
-      upvoteButton.classList.add("upvoted-arrow");
-      downvoteButton.classList.remove("downvoted-arrow");
-    } else {
-      //downvote
-      if (upvoteButton.classList.contains("upvoted-arrow")) score.innerText--;
-      downvoteButton.classList.add("downvoted-arrow");
-      upvoteButton.classList.remove("upvoted-arrow");
+    // previously voted
+    if (upOrDownCode === 1 && upvoteButton.classList.contains("upvoted-arrow")) {
+        try {
+            await fetch(`http://localhost:8080/answers/${answerId}/vote`, {
+                method: `DELETE`,
+            });
+        } catch (err) {
+            console.log("question vote error", err);
+        }
+        upvoteButton.classList.remove("upvoted-arrow");
+        score.innerText--;
+        return;
     }
-  } catch (err) {
-    console.log("question vote error", err);
-  }
+
+    // previously voted
+    if (upOrDownCode === 2 && downvoteButton.classList.contains("downvoted-arrow")) {
+        try {
+            await fetch(`http://localhost:8080/answers/${answerId}/vote`, {
+                method: `DELETE`,
+            });
+        } catch (err) {
+            console.log("question vote error", err);
+        }
+        downvoteButton.classList.remove("downvoted-arrow");
+        score.innerText++;
+        return;
+    }
+
+    // fresh vote
+    try {
+        if (downvoteButton.classList.contains("downvoted-arrow") || upvoteButton.classList.contains("upvoted-arrow")) {
+            await fetch(`http://localhost:8080/answers/${answerId}/vote`, {
+                method: `DELETE`,
+            });
+        }
+        await fetch(
+            `http://localhost:8080/answers/${answerId}/vote/${upOrDownCode}`,
+            { method: `POST` }
+        );
+
+        if (upOrDownCode === 1) { //upvote
+            if (downvoteButton.classList.contains("downvoted-arrow")) score.innerText++;
+            score.innerText++;
+            upvoteButton.classList.add("upvoted-arrow");
+            downvoteButton.classList.remove("downvoted-arrow");
+        } else { //downvote
+            if (upvoteButton.classList.contains("upvoted-arrow")) score.innerText--;
+            score.innerText--;
+            downvoteButton.classList.add("downvoted-arrow");
+            upvoteButton.classList.remove("upvoted-arrow");
+        }
+
+    } catch (err) {
+        console.log("question vote error", err);
+    }
 }
